api/requirements.txt
alembic==1.3.1
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.20.9
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.2
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
